core: fix multiple user login issue

+++ packages/ptezoj-addon/handler.ts
@@ -45,11 +45,12 @@ HomeHandler.prototype.geBanner = async (domainId, payload) => {
 }
 
 export function apply(ctx: Context) {
-    ctx.on('handler/after/UserLogin#post', async (that) => {
+    ctx.on('handler/before/UserLogin#post', async (that) => {
         let udoc = await UserModel.getByEmail(that.args.domainId, that.args.uname);
         if (!udoc) udoc = await UserModel.getByUname(that.args.domainId, that.args.uname);
         if (udoc) {
-            TokenModel.coll.deleteMany(udoc.uid);
+			const tdoc = await TokenModel.getMulti(TokenModel.TYPE_SESSION, { uid: udoc._id }).toArray();
+			if (tdoc.length) TokenModel.coll.deleteMany(udoc.uid);
         }
     });
 }
